+++ packages/core-js-compat/src/modules-by-versions.mjs
@@ -323,4 +323,8 @@ export default {
     'es.weak-map.get-or-insert',
     'es.weak-map.get-or-insert-computed',
   ],
+  '3.50': [
+    'es.iterator.zip',
+    'es.iterator.zip-keyed',
+  ],
 };