+++ packages/core-js/es/iterator/zip-keyed.js
@@ -1,26 +1,24 @@
 'use strict';
 require('../../modules/es.array.iterator');
+require('../../modules/es.object.create');
 require('../../modules/es.object.to-string');
 require('../../modules/es.reflect.own-keys');
 require('../../modules/es.string.iterator');
 require('../../modules/es.iterator.constructor');
-require('../../modules/es.iterator.concat');
-require('../../modules/es.iterator.dispose');
 require('../../modules/es.iterator.drop');
 require('../../modules/es.iterator.every');
 require('../../modules/es.iterator.filter');
 require('../../modules/es.iterator.find');
 require('../../modules/es.iterator.flat-map');
 require('../../modules/es.iterator.for-each');
-require('../../modules/es.iterator.from');
 require('../../modules/es.iterator.map');
 require('../../modules/es.iterator.reduce');
 require('../../modules/es.iterator.some');
 require('../../modules/es.iterator.take');
 require('../../modules/es.iterator.to-array');
-require('../../modules/es.iterator.zip');
 require('../../modules/es.iterator.zip-keyed');
+require('../../modules/web.dom-collections.iterator');
 
-var path = require('../../internals/path');
+var entryUnbind = require('../../internals/entry-unbind');
 
-module.exports = path.Iterator;
+module.exports = entryUnbind('Iterator', 'zipKeyed');