+++ packages/core-js/stable/iterator/zip-keyed.js
@@ -1,0 +1,5 @@
+'use strict';
+var parent = require('../../es/iterator/zip-keyed');
+require('../../modules/web.dom-collections.iterator');
+
+module.exports = parent;